% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuity.R
\docType{class}
\name{Annuity}
\alias{Annuity}
\title{Annuity}
\format{An R6 class.}
\usage{
Annuity
}
\description{
Computes present value, future value, payments and years
for annuity and annuity due.
}
\section{Arguments}{

\describe{
  \item{payment}{Annuity payment.}
  \item{rate}{Interest/discount rate.}
  \item{years}{Number of years/annuity payments.}
  \item{pv}{Present value of cash flow.}
  \item{fv}{Future value of cash flow.}
}
}

\section{Details}{


\code{$new()} will instantiate an object of class \code{Annuity}.

\code{set_payment()} will set the value for payment.

\code{set_rate()} will set the value for rate.

\code{set_years()} will set the value for years.

\code{set_pv()} will set the value for present value.

\code{set_fv()} will set the value for future value.

\code{compute_pv()} computes present value of an annuity.

\code{compute_pv_due()} computes present value of an annuity due.

\code{compute_fv()} computes future value of an annuity.

\code{compute_fv_due()} computes future value of an annuity due.

\code{compute_years()} computes years.

\code{compute_payment()} computes payments.
}

\examples{
myAnnuity <- Annuity$new()

# set values
myAnnuity$set_payment(1000)
myAnnuity$set_rate(8)
myAnnuity$set_years(5)

# compute present value of annuity and annuity due
myAnnuity$compute_pv()
myAnnuity$compute_pv_due()

# compute future value of annuity and annuity due
myAnnuity$compute_fv()
myAnnuity$compute_fv_due()

# compute payment
myAnnuity <- Annuity$new()
myAnnuity$set_rate(7.35)
myAnnuity$set_years(5)

# compute payment given present value
myAnnuity$set_pv(546.15)
myAnnuity$compute_payment()

# compute payment given future value
myAnnuity$set_fv(941.56)
myAnnuity$compute_payment(pv = FALSE)

# compute years
myAnnuity <- Annuity$new()
myAnnuity$set_rate(10)
myAnnuity$set_payment(120)

# compute years given present value
myAnnuity$set_pv(454.89)
myAnnuity$compute_years()

# compute years given future value
myAnnuity$set_fv(610.51)
myAnnuity$compute_years(pv = FALSE)

}
\keyword{datasets}
