% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cash-flow.R
\docType{class}
\name{Cashflow}
\alias{Cashflow}
\title{Cashflow}
\format{An R6 class}
\usage{
Cashflow
}
\description{
Compute present and future value of cash flow/stream, payback period,
net present value, internal rate of return, equivalent annual annuity,
profitability and break even period.
}
\section{Arguments}{

\describe{
  \item{cash_flow}{Cash flow.}
  \item{interest_rate}{Interest rate.}
  \item{discount_rate}{Discount rate.}
  \item{years}{Number of years.}
  \item{cf_stream}{Cash flow stream.}
  \item{capital_cost}{Cost of capital.}
}
}

\section{Details}{


\code{$new()} will instantiate an object of class \code{Cashflow}.

\code{set_cash_flow()} will set the value for \code{cash_flow}.

\code{set_interest_rate()} will set the value for \code{interest_rate}.

\code{set_discount_rate()} will set the value for \code{discount_rate}.

\code{set_years()} will set the value for \code{years}.

\code{set_cf_stream()} will set the value for \code{cf_stream}.

\code{set_capital_cost()} will set the value for \code{capital_cost}

\code{compute_fv()} will compute the future value of a cash flow.

\code{compute_pv()} will compute the present value of a cash flow.

\code{compute_fv_stream()} will compute the future value of a cash flow stream.

\code{compute_pv_stream()} will compute the present value of a cash flow stream.

\code{compute_payback_period()} will compute the payback period of a project.

\code{compute_npv()} will compute the net present value.

\code{compute_eaa()} will compute the equivalent annual annuity.

\code{compute_profit_index()} will compute profitability.

\code{compute_break_even()} will compute the break even period.
}

\examples{
# compute future value of a cash flow
myCashflow <- Cashflow$new()
myCashflow$set_cash_flow(100)
myCashflow$set_interest_rate(10)
myCashflow$set_years(4)
myCashflow$compute_fv()

# compute present value of a cash flow
myCashflow <- Cashflow$new()
myCashflow$set_cash_flow(39)
myCashflow$set_discount_rate(11.56)
myCashflow$set_years(7)
myCashflow$compute_pv()

# compute future value of a cash flow stream
myCashflow <- Cashflow$new()
myCashflow$set_cf_stream(c(100, 200, 200, 300))
myCashflow$set_interest_rate(10)
myCashflow$compute_fv_stream()

# compute present value of a cash flow stream
myCashflow <- Cashflow$new()
myCashflow$set_cf_stream(c(100, 200, 200, 300))
myCashflow$set_discount_rate(10)
myCashflow$compute_pv_stream()

# capital budgeting
myCashflow <- Cashflow$new()
myCashflow$set_cash_flow(c(-1000, 500, 400, 300, 200, 100))
myCashflow$set_capital_cost(10)
myCashflow$compute_payback_period()
myCashflow$compute_npv()
myCashflow$compute_eaa()
myCashflow$compute_profit_index()
myCashflow$compute_break_even()

}
\keyword{datasets}
