% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investorr.R
\docType{package}
\name{investorr}
\alias{investorr}
\alias{investorr-package}
\title{\code{investorr} package}
\description{
Tools for portfolio optimization.
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/investorr}{GitHub}
}
