% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annuities.R
\name{ivt_annuity_pv}
\alias{ivt_annuity_pv}
\alias{ivt_annuity_fv}
\alias{ivt_annuity_due_pv}
\alias{ivt_annuity_due_fv}
\alias{ivt_payment_fv}
\alias{ivt_payment_pv}
\alias{ivt_years_fv}
\alias{ivt_years_pv}
\title{Annuity}
\usage{
ivt_annuity_pv(payment, rate, periods)

ivt_annuity_fv(payment, rate, periods)

ivt_annuity_due_pv(payment, rate, periods)

ivt_annuity_due_fv(payment, rate, periods)

ivt_payment_fv(future_value, rate, periods)

ivt_payment_pv(present_value, rate, periods)

ivt_years_fv(future_value, rate, payment)

ivt_years_pv(present_value, rate, payment)
}
\arguments{
\item{payment}{Annuity payment.}

\item{future_value}{Future value of cash flow.}

\item{present_value}{Present value of cash flow.}

\item{interest_rate}{Interest rate.}

\item{discount}{Discount rate.}

\item{years}{Number of years/annuity payments.}
}
\description{
Computes present value, future value, payments and periods
for annuity and annuity due.
}
\examples{
# present value of annuity
ivt_annuity_pv(500000, 8, 5)

# future value of annuity
ivt_annuity_fv()

# present value of annuity due
ivt_annuity_due_pv()

# future value of annuity due
ivt_annuity_due_fv()

# payments given present value
ivt_payment_pv()

# payments given future value
ivt_payment_fv()

# years given present value
ivt_years_pv()

# years given future value
ivt_years_fv()

}
