% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-price.R
\name{ivt_bond_price_1}
\alias{ivt_bond_price_1}
\alias{ivt_bond_price_2}
\alias{ivt_bond_price_3}
\alias{ivt_bond_price_4}
\alias{ivt_bond_dates}
\title{Bond price}
\usage{
ivt_bond_price_1(face_value, coupon_rate, maturity_years, yield)

ivt_bond_price_2(face_value, coupon_rate, maturity_years, yield,
  coupon_frequency = c("annual", "semi-annual", "quarterly", "monthly"))

ivt_bond_price_3(par_value, redemption_value, rate, yield, settlement_date,
  maturity_date, frequency)

ivt_bond_price_4(par_value, rate, yield, settlement_date, maturity_date,
  frequency)

ivt_bond_dates(sdate, mdate, freq)
}
\arguments{
\item{face_value}{Face value of the bond.}

\item{coupon_rate}{Coupon rate of the bond.}

\item{maturity_years}{Years to maturity.}

\item{yield}{Yield to maturity.}

\item{coupon_frequency}{Frequency of coupon payments.}
}
\description{
Computes the bond price given different inputs.
}
