% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-yield.R
\name{ivt_bond_ytm}
\alias{ivt_bond_ytm}
\alias{ivt_cyield}
\alias{ivt_adj_cyield}
\title{Yield}
\usage{
ivt_bond_ytm(face_value, coupon_rate, bond_price, years)

ivt_cyield(face_value, coupon_rate, bond_price)

ivt_adj_cyield(face_value, coupon_rate, bond_price, years)
}
\arguments{
\item{face_value}{Face value of the bond.}

\item{coupon_rate}{Coupon rate of the bond.}

\item{bond_price}{Price/market value of the bond.}

\item{years}{Years to maturity.}

\item{freq}{Number of coupon payments in a year.}
}
\description{
Computes yield to maturity, yield to call, current yield and
adjusted current yield.
}
\examples{
ivt_bond_ytm(99.94, 6.25, 139.87, 12)
ivt_cyield(1000, 10, 980)
ivt_adj_cyield(100, 8, 97.26, 3)

}
