% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capm.R
\name{ivt_capm_beta}
\alias{ivt_capm_beta}
\alias{ivt_capm_stock_return}
\alias{ivt_capm_market_return}
\alias{ivt_capm_risk_free}
\title{Capital asset pricing model}
\usage{
ivt_capm_beta(risk_free, return_stock, return_market)

ivt_capm_stock_return(risk_free, return_market, beta)

ivt_capm_market_return(risk_free, return_stock, beta)

ivt_capm_risk_free(return_market, return_stock, beta)
}
\arguments{
\item{risk_free}{Risk free rate.}

\item{return_stock}{Expected return on the stock.}

\item{return_market}{Expected return on the market.}

\item{beta}{Beta for stock.}
}
\description{
Computes beta, risk free rate, expected return on
stock and market.
}
\examples{
ivt_capm_beta(1.8, 13.64, 9.2)
ivt_capm_stock_return(7.8, 10.7, 1.2)
ivt_capm_market_return(5.2, 6.5, 2.4)
ivt_capm_risk_free(7, 12.3, 2)

}
