% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf.R
\name{ivt_cf_fv}
\alias{ivt_cf_fv}
\alias{ivt_cf_pv}
\alias{ivt_cf_stream_fv}
\alias{ivt_cf_stream_pv}
\title{Cash flow}
\usage{
ivt_cf_fv(cash_flow, interest_rate, years)

ivt_cf_pv(cash_flow, discount_rate, years)

ivt_cf_stream_fv(cf_stream, interest_rate)

ivt_cf_stream_pv(cf_stream, discount_rate)
}
\arguments{
\item{cash_flow}{Cash flow.}

\item{interest_rate}{Interest rate.}

\item{years}{Number of years.}

\item{discount_rate}{Discount rate.}

\item{cf_stream}{Cash flow stream.}
}
\description{
Compute present and future value of a single cash flow and cash flow stream.
}
\examples{
# future value
ivt_cf_fv(100, 10, 4)
ivt_cf_fv(191, 10.17, 25)

# present value
ivt_cf_pv(39, 11.56, 7)
ivt_cf_pv(373, 6.67, 13)

# cash flow stream future value
ivt_cf_stream_fv(c(100, 200, 200, 300), 10)

# cash flow stream present value
ivt_cf_stream_pv(c(100, 200, 200, 300), 10)

}
