% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond-duration.R
\name{ivt_duration_macaulay}
\alias{ivt_duration_macaulay}
\alias{ivt_duration_modified}
\alias{ivt_bond_convexity}
\title{Duration}
\usage{
ivt_duration_macaulay(face_value, coupon_rate, yield, years)

ivt_duration_modified(face_value, coupon_rate, yield, years)

ivt_bond_convexity(face_value, coupon_rate, yield, years)
}
\arguments{
\item{face_value}{Face/Par value of the bond.}

\item{coupon_rate}{Coupon rate of the bond.}

\item{yield}{Yield to maturity.}

\item{years}{Years to maturity.}
}
\description{
Computes macaulay duration, modified macaulay duration and bond convexity.
}
\examples{
ivt_duration_macaulay(1000, 10, 8, 5)
ivt_duration_modified(1000, 10, 8, 5)
ivt_bond_convexity(100, 14, 12, 4)

}
