% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital-budgeting.R
\name{ivt_payback_period}
\alias{ivt_payback_period}
\alias{ivt_npv}
\alias{ivt_eaa}
\alias{ivt_profit_index}
\alias{ivt_break_even}
\title{Capital budgeting}
\usage{
ivt_payback_period(cash_flow)

ivt_npv(cash_flow, capital_cost)

ivt_eaa(cash_flow, capital_cost)

ivt_profit_index(cash_flow, capital_cost)

ivt_break_even(cash_flow)
}
\description{
Computes payback period, net present value, internal rate of return
equivalent annual annuity, profitability and break even period.
}
