% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-return.R
\name{ivt_pf_ef}
\alias{ivt_pf_ef}
\alias{plot.ivt_pf_ef}
\title{Efficient frontier}
\usage{
ivt_pf_ef(returns_1, risk_1, returns_2, risk_2, correlation)

\method{plot}{ivt_pf_ef}(x, ...)
}
\arguments{
\item{returns_1}{Returns from stock 1.}

\item{risk_1}{Risk of stock 1.}

\item{returns_2}{Returns from stock 2.}

\item{risk_2}{Risk from stock 2.}

\item{correlation}{Correlation between stock 1 and stock 2.}

\item{x}{An object of class \code{ivt_pf_ef}.}

\item{...}{Other arguments.}
}
\description{
Two stock portfolio efficient frontier.
}
\examples{
# risk and return
ivt_pf_ef(8, 10, 16, 20, -1)

# plot efficient frontier
ef <- ivt_pf_ef(8, 10, 16, 20, -1)
plot(ef)

}
