% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{ivt_stock_price}
\alias{ivt_stock_price}
\alias{ivt_stock_dividend}
\alias{ivt_stock_return}
\alias{ivt_stock_growth}
\alias{ivt_stock_ncg_price}
\title{Stock valuation}
\usage{
ivt_stock_price(dividend, growth_rate, return_rate, current = FALSE)

ivt_stock_dividend(stock_price, growth_rate, return_rate, current = FALSE)

ivt_stock_return(stock_price, dividend, growth_rate, current = FALSE)

ivt_stock_growth(stock_price, dividend, return_rate, current = FALSE)

ivt_stock_ncg_price(current_dividend, return_rate, growth_rate)
}
\arguments{
\item{dividend}{Current/next dividend.}

\item{growth_rate}{Growth rate of dividend.}

\item{return_rate}{Required return on stock.}

\item{current}{Logical; Current dividend or next dividend.}

\item{stock_price}{Price of stock.}
}
\description{
Computes the stock price, dividend, growth rate and required
rate of return.
}
\examples{
# stock price
ivt_stock_price(5.34, 1.88, 7.31)
ivt_stock_price(5.34, 1.88, 7.31, TRUE)

# stock dividend
ivt_stock_dividend(103.92, 8.6, 13.71)
ivt_stock_dividend(103.92, 8.6, 13.71, TRUE)

# stock return
ivt_stock_return(46.8, 2.41, 6.74)
ivt_stock_return(46.8, 2.41, 6.74, TRUE)

# stock growth
ivt_stock_growth(77.15, 4.44, 14.85)
ivt_stock_growth(77.15, 4.44, 14.85, TRUE)

# non constant growth
ivt_stock_ncg_price(1.22, 6.2, c(13, 7, 7, 1))

}
