% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-return.R
\name{ivt_stock_risk_return}
\alias{ivt_stock_risk_return}
\alias{ivt_pf_risk_return}
\title{Risk and return}
\usage{
ivt_stock_risk_return(probs, returns)

ivt_pf_risk_return(probs, returns_1, returns_2, correlation, weight_1)
}
\arguments{
\item{probs}{Probability distribution of future returns.}

\item{returns}{Returns on stock.}

\item{returns_1}{Returns on stock 1.}

\item{returns_2}{Returns on stock 2.}

\item{correlation}{Correlation between stock returns.}

\item{weight_1}{Proportion of portfolio invested in stock 1.}
}
\description{
Computes expected return, risk, beta and risk free rate.
}
\examples{
ivt_stock_risk_return(c(20, 30, 30, 20), c(5, 10, 15, 20))
ivt_pf_risk_return(c(20, 30, 30, 20), c(5, 10, 15, 20), c(50, 30, 10, -10),
correlation = 1, weight_1 = 0.5)

}
