% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbill.R
\name{ivt_tbill_price}
\alias{ivt_tbill_price}
\alias{ivt_tbill_rate}
\alias{ivt_rate_index}
\title{Treasury bill}
\usage{
ivt_tbill_price(rate, days)

ivt_tbill_rate(price, days)

ivt_rate_index(price, days)
}
\arguments{
\item{rate}{Annual interest rate.}

\item{days}{Days to maturity.}
}
\description{
Computes treasury bill price, rate and rate index.
}
\examples{
ivt_tbill_price(0.145, 57)
ivt_tbill_rate(93, 360)
ivt_rate_index(93, 200)

}
