% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonds.R
\name{macaulay_duration}
\alias{macaulay_duration}
\alias{macaulay_duration}
\title{Macaulay duration}
\usage{
macaulay_duration(r, i, n, M)

macaulay_duration(r, i, n, M)
}
\examples{
macaulay_duration(8, 9, 5, 1000, 4)

macaulay_duration(8, 9, 5, 1000)

}
