% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\title{Margin}
\usage{
margin(num_shares, init_price, final_price, equity, int_rate)
}
\arguments{
\item{num_shares}{Number of shares.}

\item{init_price}{Initial price of shares.}

\item{final_price}{Final price of shares.}

\item{equity}{Equity amount.}

\item{int_rate}{Interest rate.}
}
\description{
Equity margin.
}
\examples{
margin(200, 100, 70, 10000, 9)

}
