% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interest.R
\name{accrued_interest}
\alias{accrued_interest}
\title{Accrued interest}
\usage{
accrued_interest(last_date, next_date, settle_date, rate, par_value, freq)
}
\examples{
accrued_interest("2015-05-13", "2015-09-13", "2015-07-13", 7, 1000, 1)

}
