% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capital-budgeting.R
\name{ivt_payback_period}
\alias{ivt_payback_period}
\alias{ivt_npv}
\alias{ivt_eaa}
\alias{ivt_profit_index}
\alias{ivt_break_even}
\title{Capital budgeting}
\usage{
ivt_payback_period(cash_flow)

ivt_npv(cash_flow, capital_cost)

ivt_eaa(cash_flow, capital_cost)

ivt_profit_index(cash_flow, capital_cost)

ivt_break_even(cash_flow)
}
\arguments{
\item{cash_flow}{Cash flow over the project life.}

\item{capital_cost}{Cost of capital.}
}
\description{
Computes payback period, net present value, internal rate of return
equivalent annual annuity, profitability and break even period.
}
\examples{
ivt_payback_period(c(-1000, 500, 400, 300, 200, 100))
ivt_npv(c(-1000, 500, 400, 300, 200, 100), 10)
ivt_eaa(c(-1000, 500, 400, 300, 200, 100), 10)
ivt_profit_index(c(-1000, 500, 400, 300, 200, 100), 10)
ivt_break_even(c(-1000, 500, 400, 300, 200, 100))

}
